requirements.txt
flask==3.0.0
flask-cors==4.0.0
pychromecast==13.0.7
requests==2.31.0
python-crontab==3.0.0
python-dotenv==1.0.0

